refactor(web): extract custom-endpoints/onboarding mixins from web_server.py (shard s3, LateState fix)
