+++ hermes_cli/web_server.py
@@ -7104,372 +7104,24 @@ _CREDENTIAL_PROBES: dict[str, tuple[str, str]] = {
 }
 
 
-def _parse_model_ids(resp: "Any") -> List[str]:
-    """Extract model ids from an OpenAI-compatible ``/v1/models`` response.
+from hermes_cli.web_routers import providers_custom_endpoints as _providers_custom_endpoints_routes  # noqa: E402
 
-    Tolerant of the common shapes: ``{"data": [{"id": ...}]}`` (OpenAI / vLLM /
-    llama.cpp) and a bare ``{"data": ["id", ...]}``. Returns ``[]`` on any
-    parse/HTTP error so a slightly non-standard endpoint never hard-blocks.
-    """
-    try:
-        if not resp.is_success:
-            return []
-        payload = resp.json()
-    except Exception:
-        return []
-    data = payload.get("data") if isinstance(payload, dict) else payload
-    if not isinstance(data, list):
-        return []
-    ids: List[str] = []
-    for item in data:
-        if isinstance(item, dict):
-            mid = str(item.get("id") or "").strip()
-        else:
-            mid = str(item or "").strip()
-        if mid:
-            ids.append(mid)
-    return ids
-
-
-def _custom_endpoint_id(raw: str, fallback: str = "custom") -> str:
-    slug = re.sub(r"[^A-Za-z0-9_-]+", "-", (raw or "").strip()).strip("-_").lower()
-    return slug or fallback
-
-
-def _models_from_custom_endpoint_entry(entry: Dict[str, Any]) -> List[str]:
-    models: List[str] = []
-    raw_models = entry.get("models")
-    if isinstance(raw_models, dict):
-        models.extend(str(model).strip() for model in raw_models.keys())
-    elif isinstance(raw_models, list):
-        models.extend(str(model).strip() for model in raw_models)
-
-    default_model = str(entry.get("model") or entry.get("default_model") or "").strip()
-    if default_model:
-        models.insert(0, default_model)
-
-    seen: set[str] = set()
-    return [model for model in models if model and not (model in seen or seen.add(model))]
-
-
-def _api_key_display(entry: Dict[str, Any]) -> Tuple[bool, Optional[str]]:
-    """Return ``(has_api_key, preview)`` for a provider or model config block.
-
-    Keys live in ``.env`` behind ``key_env``; only entries written before
-    #69449 still carry a plaintext ``api_key``. Checking both keeps the panel
-    honest either way — reading only ``api_key`` reported "no API key" for
-    every endpoint whose key had been moved to ``.env``.
-    """
-    plaintext = str(entry.get("api_key") or "").strip()
-    if plaintext:
-        return True, redact_key(plaintext)
-    key_env = str(entry.get("key_env") or "").strip()
-    if key_env:
-        return True, f"${{{key_env}}}"
-    return False, None
-
-
-def _config_api_key_is_env_ref(endpoint_id: str) -> bool:
-    """True when this endpoint's on-disk ``api_key`` is a ``${VAR}`` template.
-
-    ``load_config()`` expands env refs, so a hand-written
-    ``api_key: ${MY_KEY}`` is indistinguishable from a literal secret by the
-    time it reaches us. Such an entry is already keeping its secret out of
-    config.yaml, so migrating it would only copy that secret into a second
-    env var the user didn't ask for.
-    """
-    providers = read_raw_config().get("providers")
-    entry = providers.get(endpoint_id) if isinstance(providers, dict) else None
-    raw_key = entry.get("api_key") if isinstance(entry, dict) else None
-    return bool(isinstance(raw_key, str) and re.search(r"\$\{[^}]+\}", raw_key))
-
-
-def _custom_endpoint_response(cfg: Dict[str, Any]) -> Dict[str, Any]:
-    model_cfg = cfg.get("model", {}) if isinstance(cfg.get("model"), dict) else {}
-    current_provider = str(model_cfg.get("provider", "") or "")
-    current_model = str(model_cfg.get("default", model_cfg.get("name", "")) or "")
-    current_base_url = str(model_cfg.get("base_url", "") or "")
-
-    endpoints: List[Dict[str, Any]] = []
-    providers = cfg.get("providers")
-    if isinstance(providers, dict):
-        for provider_id, raw_entry in providers.items():
-            if not isinstance(raw_entry, dict):
-                continue
-            base_url = str(raw_entry.get("base_url") or raw_entry.get("url") or raw_entry.get("api") or "").strip()
-            if not base_url:
-                continue
-            endpoint_id = str(provider_id)
-            models = _models_from_custom_endpoint_entry(raw_entry)
-            endpoint_model = str(raw_entry.get("model") or raw_entry.get("default_model") or (models[0] if models else ""))
-            has_api_key, api_key_preview = _api_key_display(raw_entry)
-            endpoints.append({
-                "id": endpoint_id,
-                "name": str(raw_entry.get("name") or endpoint_id),
-                "base_url": base_url,
-                "model": endpoint_model,
-                "models": models,
-                "context_length": raw_entry.get("context_length"),
-                "discover_models": bool(raw_entry.get("discover_models", True)),
-                "has_api_key": has_api_key,
-                "api_key_preview": api_key_preview,
-                "is_current": endpoint_id == current_provider,
-                "source": "providers",
-            })
-
-    if current_provider.lower() == "custom" and current_base_url and not any(e["id"] == "custom" for e in endpoints):
-        has_api_key, api_key_preview = _api_key_display(model_cfg)
-        endpoints.insert(0, {
-            "id": "custom",
-            "name": "Custom",
-            "base_url": current_base_url,
-            "model": current_model,
-            "models": [current_model] if current_model else [],
-            "context_length": model_cfg.get("context_length"),
-            "discover_models": True,
-            "has_api_key": has_api_key,
-            "api_key_preview": api_key_preview,
-            "is_current": True,
-            "source": "direct-config",
-        })
-
-    return {
-        "endpoints": endpoints,
-        "current": {
-            "provider": current_provider,
-            "model": current_model,
-            "base_url": current_base_url,
-        },
-    }
-
-
-def _detach_main_model_from_provider(cfg: Dict[str, Any], provider_key: str) -> None:
-    """Drop the main-slot mirror of a provider that no longer exists.
-
-    ``activate_custom_endpoint`` copies the endpoint's ``base_url`` and
-    ``api_key`` onto ``model``. That mirror outranks the environment at client
-    construction (#62269), so deleting the endpoint without clearing it leaves
-    the agent still authenticating to the deleted host with the deleted key —
-    and leaves that key sitting in config.yaml after the operator believes the
-    dashboard removed it.
-
-    Only touches ``model`` when it actually names the deleted provider, so an
-    endpoint deleted while a *different* provider is active is left alone.
-    """
-    model_cfg = cfg.get("model")
-    if not isinstance(model_cfg, dict):
-        return
-    if str(model_cfg.get("provider") or "").strip().lower() != provider_key:
-        return
-    for field in ("provider", "base_url", "api_key", "key_env"):
-        model_cfg.pop(field, None)
-    cfg["model"] = model_cfg
-
-
-def _write_custom_endpoint(cfg: Dict[str, Any], body: CustomEndpointUpdate) -> Tuple[str, Dict[str, Any]]:
-    endpoint_id = _custom_endpoint_id(body.id or body.name)
-    name = (body.name or "").strip()
-    base_url = (body.base_url or "").strip().rstrip("/")
-    model = (body.model or "").strip()
-
-    if not name:
-        raise HTTPException(status_code=400, detail="name required")
-    if not base_url:
-        raise HTTPException(status_code=400, detail="base_url required")
-    parsed = urllib.parse.urlparse(base_url)
-    if not parsed.scheme or not parsed.netloc:
-        raise HTTPException(status_code=400, detail="base_url must include scheme and host")
-    if not model:
-        raise HTTPException(status_code=400, detail="model required")
-
-    providers = cfg.get("providers")
-    if not isinstance(providers, dict):
-        providers = {}
-    existing = providers.get(endpoint_id)
-    if not isinstance(existing, dict):
-        existing = {}
-
-    # Merge onto the existing entry rather than replacing it. A providers.<name>
-    # block is not owned by this panel: it can carry hand-written keys the
-    # dashboard has no field for — ``api_mode``, ``key_env``/``api_key_env``,
-    # ``extra_headers`` (which may themselves carry credentials),
-    # ``request_overrides`` — and rebuilding from scratch silently dropped every
-    # one of them on an unrelated edit, leaving a provider that no longer
-    # authenticates or speaks the right protocol.
-    entry: Dict[str, Any] = dict(existing)
-    entry.update({
-        "name": name,
-        "base_url": base_url,
-        "model": model,
-        "discover_models": bool(body.discover_models),
-    })
-    # Same for the model map: merge rather than replace, so existing models
-    # keep their context lengths. ``body.models`` is the catalogue the panel's
-    # Test button already discovered — without it only the one hand-typed
-    # model survived Save, and every picker showed a single-entry list for a
-    # provider serving dozens (#69988). A payload with no ``models`` (older
-    # UI) still just ensures the named default is present.
-    existing_models = entry.get("models")
-    models_map: Dict[str, Any] = dict(existing_models) if isinstance(existing_models, dict) else {}
-    for candidate in (*(body.models or ()), model):
-        model_id = str(candidate).strip()
-        if not model_id:
-            continue
-        current = models_map.get(model_id)
-        models_map[model_id] = dict(current) if isinstance(current, dict) else {}
-    entry["models"] = models_map
-    if body.context_length and body.context_length > 0:
-        entry["context_length"] = int(body.context_length)
-        entry["models"][model]["context_length"] = int(body.context_length)
-
-    # API keys never belong in config.yaml (#69449). Write to .env and
-    # reference it via ``key_env`` — the same indirection built-in providers
-    # use and that runtime_provider.py already resolves at load time.
-    env_var = custom_endpoint_key_env(endpoint_id)
-    submitted_key = body.api_key.strip() if body.api_key is not None else None
-    if submitted_key:
-        save_env_value(env_var, submitted_key)
-        entry["key_env"] = env_var
-        entry.pop("api_key", None)
-    elif submitted_key is not None:
-        # Blank field means "clear the key", not "leave it alone".
-        remove_env_value(env_var)
-        entry.pop("key_env", None)
-        entry.pop("api_key", None)
-    elif str(entry.get("api_key") or "").strip() and not _config_api_key_is_env_ref(endpoint_id):
-        # No new key submitted, but this entry still carries one an earlier
-        # release wrote in plaintext. Migrate it on the next save so endpoints
-        # configured before the fix get cleaned up too, without the user
-        # having to re-enter the key.
-        save_env_value(env_var, entry["api_key"].strip())
-        entry["key_env"] = env_var
-        entry.pop("api_key", None)
-
-    providers[endpoint_id] = entry
-    cfg["providers"] = providers
-
-    if body.make_default:
-        cfg["model"] = _apply_main_model_assignment(
-            cfg.get("model", {}), endpoint_id, model, base_url
-        )
-        if entry.get("key_env") and isinstance(cfg["model"], dict):
-            cfg["model"]["key_env"] = entry["key_env"]
-            cfg["model"].pop("api_key", None)
-
-    return endpoint_id, entry
-
-
-@app.get("/api/providers/custom-endpoints")
-def list_custom_endpoints():
-    """Return configured OpenAI-compatible custom endpoints for Desktop."""
-    try:
-        return _custom_endpoint_response(load_config())
-    except Exception:
-        _log.exception("GET /api/providers/custom-endpoints failed")
-        raise HTTPException(status_code=500, detail="Failed to list custom endpoints")
-
-
-@app.post("/api/providers/custom-endpoints")
-def upsert_custom_endpoint(body: CustomEndpointUpdate):
-    """Create or update a v12+ ``providers`` custom endpoint entry."""
-    try:
-        cfg = load_config()
-        endpoint_id, _entry = _write_custom_endpoint(cfg, body)
-        save_config(cfg)
-        response = _custom_endpoint_response(cfg)
-        response["ok"] = True
-        response["id"] = endpoint_id
-        return response
-    except HTTPException:
-        raise
-    except Exception:
-        _log.exception("POST /api/providers/custom-endpoints failed")
-        raise HTTPException(status_code=500, detail="Failed to save custom endpoint")
-
-
-@app.post("/api/providers/custom-endpoints/{endpoint_id}/activate")
-def activate_custom_endpoint(endpoint_id: str):
-    """Set a configured custom endpoint as the default model provider."""
-    try:
-        cfg = load_config()
-        provider_key = _custom_endpoint_id(endpoint_id)
-        providers = cfg.get("providers")
-        entry = providers.get(provider_key) if isinstance(providers, dict) else None
-        if not isinstance(entry, dict):
-            raise HTTPException(status_code=404, detail="custom endpoint not found")
-
-        models = _models_from_custom_endpoint_entry(entry)
-        model = str(entry.get("model") or (models[0] if models else "")).strip()
-        base_url = str(entry.get("base_url") or "").strip()
-        if not model or not base_url:
-            raise HTTPException(status_code=400, detail="custom endpoint is incomplete")
-
-        model_cfg = _apply_main_model_assignment(cfg.get("model", {}), provider_key, model, base_url)
-        if entry.get("key_env"):
-            model_cfg["key_env"] = entry["key_env"]
-            model_cfg.pop("api_key", None)
-        elif entry.get("api_key"):
-            model_cfg["api_key"] = entry["api_key"]
-        cfg["model"] = model_cfg
-        save_config(cfg)
-        return {"ok": True, "provider": provider_key, "model": model}
-    except HTTPException:
-        raise
-    except Exception:
-        _log.exception("POST /api/providers/custom-endpoints/%s/activate failed", endpoint_id)
-        raise HTTPException(status_code=500, detail="Failed to activate custom endpoint")
-
-
-@app.delete("/api/providers/custom-endpoints/{endpoint_id}")
-def delete_custom_endpoint(endpoint_id: str):
-    """Remove a configured custom endpoint from ``providers``."""
-    try:
-        cfg = load_config()
-        provider_key = _custom_endpoint_id(endpoint_id)
-        providers = cfg.get("providers")
-        if not isinstance(providers, dict) or provider_key not in providers:
-            raise HTTPException(status_code=404, detail="custom endpoint not found")
-        providers.pop(provider_key, None)
-        cfg["providers"] = providers
-        _detach_main_model_from_provider(cfg, provider_key)
-        remove_env_value(custom_endpoint_key_env(provider_key))
-        save_config(cfg)
-        response = _custom_endpoint_response(cfg)
-        response["ok"] = True
-        return response
-    except HTTPException:
-        raise
-    except Exception:
-        _log.exception("DELETE /api/providers/custom-endpoints/%s failed", endpoint_id)
-        raise HTTPException(status_code=500, detail="Failed to delete custom endpoint")
-
-
-@app.post("/api/providers/custom-endpoints/validate")
-async def validate_custom_endpoint(body: CustomEndpointUpdate):
-    """Probe a custom endpoint by calling its OpenAI-compatible /models URL."""
-    import httpx
-
-    base_url = (body.base_url or "").strip().rstrip("/")
-    if not base_url:
-        return {"ok": False, "reachable": True, "message": "Enter an endpoint URL first.", "models": []}
-
-    url = base_url + "/models"
-    headers = {"Accept": "application/json"}
-    if body.api_key and body.api_key.strip():
-        headers["Authorization"] = f"Bearer {body.api_key.strip()}"
-
-    try:
-        async with httpx.AsyncClient(timeout=httpx.Timeout(8.0)) as client:
-            resp = await client.get(url, headers=headers)
-    except Exception:
-        return {"ok": False, "reachable": False, "message": f"Could not reach {url}.", "models": []}
-
-    if resp.status_code in (401, 403):
-        return {"ok": False, "reachable": True, "message": "The endpoint rejected the API key.", "models": []}
-    if not resp.is_success:
-        return {"ok": False, "reachable": True, "message": f"Endpoint returned HTTP {resp.status_code}.", "models": []}
-
-    return {"ok": True, "reachable": True, "message": "", "models": _parse_model_ids(resp)}
+app.include_router(_providers_custom_endpoints_routes.router)
+from hermes_cli.web_routers.providers_custom_endpoints import (  # noqa: E402,F401 — legacy re-exports; tests call these via web_server.<name>
+    _api_key_display,
+    _config_api_key_is_env_ref,
+    _custom_endpoint_id,
+    _custom_endpoint_response,
+    _detach_main_model_from_provider,
+    _models_from_custom_endpoint_entry,
+    _parse_model_ids,
+    _write_custom_endpoint,
+    activate_custom_endpoint,
+    delete_custom_endpoint,
+    list_custom_endpoints,
+    upsert_custom_endpoint,
+    validate_custom_endpoint,
+)
 
 
 @app.post("/api/providers/validate")
@@ -8418,70 +8070,6 @@ from hermes_cli.web_routers.whatsapp_onboarding import (  # noqa: E402,F401 — 
     get_whatsapp_onboarding_status,
     start_whatsapp_onboarding,
 )
-_TELEGRAM_ONBOARDING_DEFAULT_URL = "https://setup.hermes-agent.nousresearch.com"
-_TELEGRAM_ONBOARDING_USER_AGENT = f"HermesDashboard/{__version__}"
-@dataclass
-class _TelegramOnboardingPairing:
-    poll_token: str
-    expires_at: str
-    expires_at_ts: float
-    bot_token: str | None = None
-    bot_username: str | None = None
-    owner_user_id: str | None = None
-
-
-_telegram_onboarding_pairings: dict[str, _TelegramOnboardingPairing] = {}
-_telegram_onboarding_lock = threading.RLock()
-
-
-def _telegram_onboarding_base_url() -> str:
-    return (
-        os.getenv("TELEGRAM_ONBOARDING_URL", _TELEGRAM_ONBOARDING_DEFAULT_URL)
-        .strip()
-        .rstrip("/")
-    )
-
-
-def _parse_expiry_ts(value: str) -> float:
-    try:
-        normalized = value.replace("Z", "+00:00")
-        parsed = datetime.fromisoformat(normalized)
-        if parsed.tzinfo is None:
-            parsed = parsed.replace(tzinfo=timezone.utc)
-        return parsed.timestamp()
-    except Exception:
-        return time.time() + 600
-
-
-def _prune_telegram_onboarding_pairings() -> None:
-    now = time.time()
-    expired = [
-        pairing_id
-        for pairing_id, record in _telegram_onboarding_pairings.items()
-        if record.expires_at_ts <= now
-    ]
-    for pairing_id in expired:
-        _telegram_onboarding_pairings.pop(pairing_id, None)
-
-
-def _normalize_telegram_user_id(value: Any) -> str | None:
-    normalized = str(value or "").strip()
-    if _TELEGRAM_USER_ID_RE.fullmatch(normalized):
-        return normalized
-    return None
-
-
-def _telegram_onboarding_error_message(error: str, fallback: str) -> str:
-    return {
-        "not_found": "Telegram pairing was not found. Start a new setup.",
-        "expired": "Telegram setup expired. Start a new setup.",
-        "claimed": "Telegram setup was already claimed. Start a new setup.",
-        "unauthorized": "Telegram setup service rejected this request.",
-        "telegram_manager_bot_token_not_configured": "Telegram setup service is not configured.",
-        "telegram_token_fetch_failed": "Telegram could not finish bot setup. Try again.",
-    }.get(error, fallback)
-
-
 def _telegram_onboarding_request_sync(
     method: str,
     path: str,
@@ -8552,236 +8140,27 @@ def _telegram_onboarding_request_sync(
     return parsed
 
 
-async def _telegram_onboarding_request(
-    method: str,
-    path: str,
-    *,
-    body: dict[str, Any] | None = None,
-    bearer_token: str | None = None,
-) -> dict[str, Any]:
-    return await asyncio.to_thread(
-        _telegram_onboarding_request_sync,
-        method,
-        path,
-        body=body,
-        bearer_token=bearer_token,
-    )
+from hermes_cli.web_routers import telegram_onboarding as _telegram_onboarding_routes  # noqa: E402
 
-
-@app.post("/api/messaging/telegram/onboarding/start")
-async def start_telegram_onboarding(body: TelegramOnboardingStart):
-    bot_name = (body.bot_name or "Hermes Agent").strip() or "Hermes Agent"
-    payload = await _telegram_onboarding_request(
-        "POST",
-        "/v1/telegram/pairings",
-        body={"bot_name": bot_name},
-    )
-
-    pairing_id = str(payload.get("pairing_id") or "").strip()
-    poll_token = str(payload.get("poll_token") or "").strip()
-    expires_at = str(payload.get("expires_at") or "").strip()
-    deep_link = str(payload.get("deep_link") or "").strip()
-    qr_payload = str(payload.get("qr_payload") or deep_link).strip()
-    suggested_username = str(payload.get("suggested_username") or "").strip()
-    if not pairing_id or not poll_token or not expires_at or not deep_link:
-        raise HTTPException(
-            status_code=502,
-            detail="Telegram setup service returned an incomplete response.",
-        )
-
-    with _telegram_onboarding_lock:
-        _prune_telegram_onboarding_pairings()
-        _telegram_onboarding_pairings[pairing_id] = _TelegramOnboardingPairing(
-            poll_token=poll_token,
-            expires_at=expires_at,
-            expires_at_ts=_parse_expiry_ts(expires_at),
-        )
-
-    return {
-        "pairing_id": pairing_id,
-        "suggested_username": suggested_username,
-        "deep_link": deep_link,
-        "qr_payload": qr_payload,
-        "expires_at": expires_at,
-    }
-
-
-@app.get("/api/messaging/telegram/onboarding/{pairing_id}")
-async def get_telegram_onboarding_status(pairing_id: str):
-    with _telegram_onboarding_lock:
-        _prune_telegram_onboarding_pairings()
-        record = _telegram_onboarding_pairings.get(pairing_id)
-        if not record:
-            raise HTTPException(
-                status_code=404,
-                detail="Telegram setup session was not found. Start a new setup.",
-            )
-        if record.bot_token:
-            return {
-                "status": "ready",
-                "bot_username": record.bot_username,
-                "owner_user_id": record.owner_user_id,
-                "expires_at": record.expires_at,
-            }
-        poll_token = record.poll_token
-
-    payload = await _telegram_onboarding_request(
-        "GET",
-        f"/v1/telegram/pairings/{urllib.parse.quote(pairing_id, safe='')}",
-        bearer_token=poll_token,
-    )
-    status = str(payload.get("status") or "").strip()
-    if status == "waiting":
-        with _telegram_onboarding_lock:
-            current = _telegram_onboarding_pairings.get(pairing_id)
-            expires_at = current.expires_at if current else ""
-        return {"status": "waiting", "expires_at": expires_at}
-
-    if status == "ready":
-        bot_token = str(payload.get("token") or "").strip()
-        bot_username = str(payload.get("bot_username") or "").strip()
-        if not bot_token:
-            raise HTTPException(
-                status_code=502,
-                detail="Telegram setup service returned an incomplete response.",
-            )
-        owner_user_id = _normalize_telegram_user_id(payload.get("owner_user_id"))
-        with _telegram_onboarding_lock:
-            record = _telegram_onboarding_pairings.get(pairing_id)
-            if not record:
-                raise HTTPException(
-                    status_code=404,
-                    detail="Telegram setup session was not found. Start a new setup.",
-                )
-            record.bot_token = bot_token
-            record.bot_username = bot_username or None
-            record.owner_user_id = owner_user_id
-            return {
-                "status": "ready",
-                "bot_username": record.bot_username,
-                "owner_user_id": record.owner_user_id,
-                "expires_at": record.expires_at,
-            }
-
-    if status in {"expired", "claimed"}:
-        with _telegram_onboarding_lock:
-            _telegram_onboarding_pairings.pop(pairing_id, None)
-        raise HTTPException(
-            status_code=410,
-            detail=_telegram_onboarding_error_message(
-                status,
-                "Telegram setup is no longer available. Start a new setup.",
-            ),
-        )
-
-    raise HTTPException(
-        status_code=502,
-        detail="Telegram setup service returned an unknown status.",
-    )
-
-
-def _restart_gateway_after_telegram_onboarding(profile: Optional[str] = None) -> dict[str, Any]:
-    """Best-effort gateway restart after saving Telegram QR onboarding.
-
-    The QR flow naturally pulls users into Telegram on another device. If the
-    saved token waits on a separate dashboard restart click, Hermes appears
-    broken from the chat side. Keep the config save authoritative, but report
-    restart failures so the UI can fall back to the existing manual banner.
-    """
-    try:
-        proc, reused = _spawn_gateway_restart(profile)
-    except Exception as exc:
-        _log.exception("Failed to auto-restart gateway after Telegram onboarding")
-        return {
-            "restart_started": False,
-            "restart_error": str(exc),
-        }
-    if reused:
-        _log.info(
-            "Telegram onboarding: reusing in-flight gateway restart (pid %s)",
-            proc.pid,
-        )
-    return {
-        "restart_started": True,
-        "restart_action": "gateway-restart",
-        "restart_pid": proc.pid,
-    }
-
-
-@app.post("/api/messaging/telegram/onboarding/{pairing_id}/apply")
-async def apply_telegram_onboarding(
-    pairing_id: str, body: TelegramOnboardingApply, profile: Optional[str] = None
-):
-    allowed_user_ids = []
-    seen = set()
-    for raw_id in body.allowed_user_ids:
-        normalized = _normalize_telegram_user_id(raw_id)
-        if not normalized:
-            raise HTTPException(
-                status_code=400,
-                detail="Allowed Telegram user IDs must be numeric.",
-            )
-        if normalized not in seen:
-            seen.add(normalized)
-            allowed_user_ids.append(normalized)
-    if not allowed_user_ids:
-        raise HTTPException(
-            status_code=400,
-            detail="Add at least one allowed Telegram user ID.",
-        )
-
-    with _telegram_onboarding_lock:
-        _prune_telegram_onboarding_pairings()
-        record = _telegram_onboarding_pairings.get(pairing_id)
-        if not record:
-            raise HTTPException(
-                status_code=404,
-                detail="Telegram setup session was not found. Start a new setup.",
-            )
-        bot_token = record.bot_token
-        bot_username = record.bot_username
-        if not bot_token:
-            raise HTTPException(
-                status_code=409,
-                detail="Telegram setup is not ready yet.",
-            )
-
-    effective_profile = body.profile or profile
-    try:
-        with _profile_scope(effective_profile):
-            save_env_value("TELEGRAM_BOT_TOKEN", bot_token)
-            save_env_value("TELEGRAM_ALLOWED_USERS", ",".join(allowed_user_ids))
-            _write_platform_enabled("telegram", True)
-    except HTTPException:
-        raise
-    except ValueError as exc:
-        raise HTTPException(status_code=400, detail=str(exc)) from exc
-    except Exception as exc:
-        _log.exception("Telegram onboarding apply failed")
-        raise HTTPException(
-            status_code=500,
-            detail="Failed to save Telegram setup.",
-        ) from exc
-
-    with _telegram_onboarding_lock:
-        _telegram_onboarding_pairings.pop(pairing_id, None)
-
-    restart_result = _restart_gateway_after_telegram_onboarding(effective_profile)
-
-    return {
-        "ok": True,
-        "platform": "telegram",
-        "bot_username": bot_username,
-        "needs_restart": not restart_result["restart_started"],
-        **restart_result,
-    }
-
-
-@app.delete("/api/messaging/telegram/onboarding/{pairing_id}")
-async def cancel_telegram_onboarding(pairing_id: str):
-    with _telegram_onboarding_lock:
-        _telegram_onboarding_pairings.pop(pairing_id, None)
-    return {"ok": True}
+app.include_router(_telegram_onboarding_routes.router)
+from hermes_cli.web_routers.telegram_onboarding import (  # noqa: E402,F401 — legacy re-exports; tests call these via web_server.<name>
+    _TELEGRAM_ONBOARDING_DEFAULT_URL,
+    _TELEGRAM_ONBOARDING_USER_AGENT,
+    _TelegramOnboardingPairing,
+    _normalize_telegram_user_id,
+    _parse_expiry_ts,
+    _prune_telegram_onboarding_pairings,
+    _restart_gateway_after_telegram_onboarding,
+    _telegram_onboarding_base_url,
+    _telegram_onboarding_error_message,
+    _telegram_onboarding_lock,
+    _telegram_onboarding_pairings,
+    _telegram_onboarding_request,
+    apply_telegram_onboarding,
+    cancel_telegram_onboarding,
+    get_telegram_onboarding_status,
+    start_telegram_onboarding,
+)
 
 
 @app.get("/api/messaging/platforms")
